backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djoser==2.2.3
psycopg2-binary==2.9.10
django-filter==24.3
Pillow==11.0.0
reportlab==4.2.5
PyYAML==6.0.2